requirements.txt
Django==4.2
djangorestframework
psycopg2-binary
python-dotenv
django-filter
django-cors-headers
djangorestframework-simplejwt
Pillow
python-magic
django-cleanup
drf-yasg
gunicorn